Added PermissionToken.perform Ref.orIfNone, tests.
Ref.orIfNone will return the passed-in default only if it is RefNone
(i.e., not if it is RefFailed). Note that Ref.toOption.orElse will
return the passed-in default for RefNone and RefFailed.
PermissionToken.perform will perform a function (a by-name param) if
permission is allowed, or return a RefFailed if permission was refused.

The usefulness of these comes from being able to use ResolvedRef like
the Lift framework's Box class -- it is something, nothing or an error.
 UnresolvedRef is a reference that becomes one of these when it is
looked up.