Added PermissionToken.check
PermissionToken.check will check if a permission would be allowable,
caching any successes, but will not fall into a refused state if the
permission was not granted. This is useful where permissions are
being checked in advance -- for instance when checking to see whether
an "edit" link should be shown (only show if the edit request would
be approved, but refusing permission should not cause other checks we
make to be pre-refused).