Experimental reworking
There's a lot of tinkering gone into this changeset, so it's hard to
provide a concise description.

Ref.gettId should no longer require reflection, but uses implicits to
be typesafe.

Ref.flatMap allows some subclasses to narrow the return type to R[B]
(the return type of the function passed in)

The TraversableOnce trait has been moved down to ResolvedRef rather
than Ref, as it may be hard for some Refs (such as an asynchronous
RefEnumerator) to implement