Supported DataAction.returning.one notation
DataAction is a case class, so that the implicit UserFromRequest is resolved before the implicit JsonConverter. (Otherwise the scala compiler won't manage to automatically infer the type of request in the block).

As this is a two step process, I've aliased DataAction.returning to the constructor.  This then means that the following methods are available for actions:

DataAction.returning.one
DataAction.returning.many
DataAction.returning.result

(which reads rather neatly)