Added mechanism to set headers in DataAction.returning methods (partial commit -- see TODO)
DataAction.returning methods now expect the block to return a WithHeaders(Ref[T], Ref[HeaderInfo])

If the block just returns a Ref[T], it is implicitly converted to WithHeaders(data, headerInfo=RefNone)

TODO: Apply headers even if the data fails.
Because the data and headerInfo are in separate Refs, headers could be applied even if the data result is a failure. However, this commit does not do that yet.