Breaking change: Identity.value is optional
Although the id / value should always be there when the social login arrives, there are times you might want an identity record in the database that doesn't have an id. For instance, GitHub has changeable usernames, but the username not the id is typically what people will give you (e.g., if you're pre-creating users from a CSV file). In which case it can be useful to place the Identity in the database with the username, and then fill in the id when the user logs in via OAuth.