Answer holds a var, for ease of use from Scala.js
On the server-side, where requests happen in parallel, we tend to want to use these immutably. However, in Scala.js everything is single-threaded and there are occasions when just being able to modify a var is helpful.

Particularly, this is useful in forms, where we may want to pass a sequence of answers to a sequence of controls, and allow the control to modify the answer rather than need to deal with Lenses or copying the whole datastructure.

Later, this change might be reflected to the other app data types, to allow them to be used more simply on the client side.