Added Latch
A latch is a clearable future, that can chain together as a "lazy observable". That is, while it has a value, it listens for changes in its parent, but it's only action is to clear itself (and deregister its listener) if its parent changes -- it waits for the next request for its own value to re-establish itself.