Added Latch, RefCache, some testing on streams
This is a belated commit to preserve some work that was done on handy some time ago that was left uncommitted on my computer. It will need a good clean up because it has been so long I have forgotten which parts were ready. Committed as I'm about to make a 0.9 branch which will revisit some of the core design and will involve some clean-up work anyway.