Fix case where re-subscribing to MapR causes hang
This is not a thread-safe fix, but does seem to solve the common case. It needs looking at later.
In MapR, if a new subscriber subscribed after the stream was finished, they would be sent a subscription that would never receive any events.
This change causes MapR to forget that it is subscribed, and clear its record of downstream subscribers, when it receives a complete signal from upstream. This then lets new subscriptions cause it to re-subscribe upstream.