Overload flatMap
Previously, we were using RCFMT (Ref Can Flatmap To) to direct flatMap calls to the correct version (flatMapOne, flatMapOpt, flatMapMany).
However, LazyId[T,K] was causing issues with this, as the compiler had changed behaviour between 2.12 and 2.13, and was effectively forcing the function to produce LazyId[T, Any], rather than LazyId[T, K] for any K

RCFMT was a similar strategy to the Scala 2.12 collections' CanBuildFrom. Scala 2.13's collections have moved away from this towards overloading.

So, this commit introduces overloaded versions of flatMap. They pick up a single implicit parameter, for which there is only one value (already provided in the companion object). This gets us past the problem that type erasure means Function[A, Ref[B]] and Function[A, RefOpt[B]] are the same at runtime and so the function itself cannot be used as the parameter to distinguish between the overloaded functions. Instead, we force the compiler to insert a second implicit argument that the runtime then uses when resolving the overloading.

The old RCFMT objects still remain present in this commit because leaving them there appears not to break anything yet, but they might be removed in future.