Partial commit. Add initial draft of Task
A Task is a function that, given an ExecutionContext, will produce a Ref.
TaskOpt and TaskMany are the equivalent for RefOpt and RefMany.

The commit also contains a set of extension methods that allow Task, TaskOpt, and TaskMany to be used in for-comprehensions and the pluralities to come out ok