Fix bug where Ref.withFilter was producing Ref despite comment
Ref.withFilter should produce RefOpt according to the comment

This is so the false case does not produce a RefFailed which might be unexpected in for-comprehensions. eg.

val ri = RefIterableRef(Seq(
  1.itself, 2.itself, 3.itself, 4.itself, 5.itself
))

(for n <- ri if n % 2 == 0 yield n).collect

In this case, we'd like to get a Ref(Seq(2, 4)) rather than accidentally get a RefFailed because the if case returned false on a Ref (rather than a RefOpt)